🍱 Add: 본인에게 팔로우와 언팔로우 예외처리 추가 (#68)

+++ src/main/kotlin/com/eatsfinder/domain/follow/service/FollowServiceImpl.kt
@@ -3,6 +3,7 @@ package com.eatsfinder.domain.follow.service
 import com.eatsfinder.domain.follow.model.Follow
 import com.eatsfinder.domain.follow.repository.FollowRepository
 import com.eatsfinder.domain.user.repository.UserRepository
+import com.eatsfinder.global.exception.InvalidInputException
 import com.eatsfinder.global.exception.ModelNotFoundException
 import com.eatsfinder.global.exception.like.DefaultZeroException
 import org.springframework.stereotype.Service
@@ -28,6 +29,8 @@ class FollowServiceImpl(
 
         if (user.followerCount <= 0 && followingUser.followingCount <= 0) throw DefaultZeroException("사용자의 팔로워 수 또는 팔로잉 수가 0 이하입니다.")
 
+        if (userId == followingUserId) throw InvalidInputException("본인에게 팔로우를 할 수는 없습니다.")
+
         if (follow == null) {
             followRepository.save(
                 Follow(
@@ -40,7 +43,7 @@ class FollowServiceImpl(
             userRepository.save(user)
             userRepository.save(followingUser)
         } else {
-            throw ModelNotFoundException("follow", "팔로우(${follow})는 한번 밖에 하지 못합니다.")
+            throw ModelNotFoundException("follow", "이미 팔로우(${follow})중입니다.")
         }
     }
 
@@ -60,6 +63,8 @@ class FollowServiceImpl(
 
         if (user.followerCount <= 0 && unfollowingUser.followingCount <= 0) throw DefaultZeroException("사용자의 팔로워 수 또는 팔로잉 수가 0 이하입니다.")
 
+        if (userId == unfollowingUserId) throw InvalidInputException("본인에게 언팔로우를 할 수 없습니다.")
+
         if (follow != null) {
             followRepository.delete(follow)
             user.followingCount--
@@ -67,7 +72,7 @@ class FollowServiceImpl(
             userRepository.save(user)
             userRepository.save(unfollowingUser)
         } else {
-            throw ModelNotFoundException("follow", "팔로우(${follow})는 한번 밖에 취소하지 못합니다.")
+            throw ModelNotFoundException("follow", "이미 언팔로우(${follow})하셨습니다.")
         }
     }
 }
